imu and qtr removed

+++ Sensorik_Aktorik/aktorik.cpp
@@ -13,7 +13,7 @@ ros::Subscriber<std_msgs::UInt8> sub_light_b("lichtBremse", lichtBremse_cb);
 ros::Subscriber<std_msgs::UInt8> sub_light_rem("lichtRemote", lichtRemote_cb);
 
 short rc_mode = 0;
-int8_t analogvalue_1; //eingelesener Pin - Wert am Tiefpass
+int16_t analogvalue_1; //eingelesener Pin - Wert am Tiefpass
   
 float wert_servo;
 float lenkwinkel_servosize;
@@ -95,7 +95,7 @@ void servo_cb(const std_msgs::Float32& cmd_msg){   //Callback - Funktion für Se
   servo_bewegung(cmd_msg.data);
 }
 
-void motor_cb(const std_msgs::UInt16& cmd_msg){     //Callback - Funktion für Motoraufruf
+void motor_cb(const std_msgs::Int16& cmd_msg){     //Callback - Funktion für Motoraufruf
   motor_bewegung(cmd_msg.data);
 }
 
@@ -125,7 +125,7 @@ void servo_bewegung(float lenkwinkel_bogenmass){
   servo.write(lenkwinkel_servosize); //Servo fährt in die ensprechende Stellung
 }
 
-void motor_bewegung(int8_t motor_drehzahl){
+void motor_bewegung(int16_t motor_drehzahl){
   int8_t motor_uebertragung;
   if(motor_drehzahl < 0){     //rückwärts
     motor_uebertragung = 90 +(0.234 * motor_drehzahl);
@@ -139,7 +139,7 @@ void motor_bewegung(int8_t motor_drehzahl){
  motor.write(motor_uebertragung);
 }
 
-void motor_bewegung_RC_mode(int8_t voltage_1){
+void motor_bewegung_RC_mode(int16_t voltage_1){
   int8_t motor_uebertragung_RC_mode;
     if (voltage_1 < tiefpass_untere_spannung){        //rückwarts
         motor_uebertragung_RC_mode = (voltage_1 * 135.23) + 56.47;
